backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
aiosqlite==0.20.0
sse-starlette==2.1.0
pydantic==2.7.1
python-multipart==0.0.9
python-dotenv==1.0.1
fpdf2==2.7.9
qrcode[pil]==7.4.2
PyMuPDF==1.24.3
websockets==12.0
Pillow==10.3.0
